feat(orders): dziesiętna stawka klienta + nullable contract start_date (import Polkomtel) (#594)
* feat(orders): dziesiętna stawka klienta (Numeric 10,2) + nullable contract start_date

Import zamówień od klienta (np. Polkomtel) gdzie stawka klienta bywa
dziesiętna (118.13, 218.75 PLN/h), a kolumna 'Zamówienie od' jest pusta.

- migracja 0147: client_orders.rate_client Integer→Numeric(10,2),
  contracts.start_date NOT NULL→nullable
- ClientOrder.rate_client + schematy zamówień (rate_client/marża) → Decimal
- create_contract_with_order: Order trzyma dokładny Decimal; Contract.rate_client
  (kolumna int) dostaje zaokrąglenie ROUND_HALF_UP; daty start opcjonalne
- read-schematy kontraktów/profilu klienta: start_date wymagane→Optional


* style: ruff format client_order schema (single-line Field)

---------

+++ backend/app/models/client_order.py
@@ -91,7 +91,10 @@ class ClientOrder(Base, TimestampMixin):
     # Rate_client per Order — może różnić się od Contract.rate_client przy
     # przedłużeniach z podwyżką. rate_candidate trzymamy na Contract (typically
     # stała przez całą współpracę z kontraktorem).
-    rate_client: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    # Numeric(10,2) — stawka klienta z PO może być dziesiętna (np. 118.13 PLN/h).
+    rate_client: Mapped[Optional[Decimal]] = mapped_column(
+        Numeric(10, 2), nullable=True
+    )
 
     total_value: Mapped[Optional[Decimal]] = mapped_column(
         Numeric(12, 2), nullable=True
